continued_articles

+++ database/migrations/2021_10_20_161844_create_articles_table.php
@@ -16,7 +16,7 @@ class CreateArticlesTable extends Migration
         Schema::create('articles', function (Blueprint $table) {
             $table->id();
             $table->string('title',155);
-            $table->integer('image_id');
+            $table->bigInteger('image_id'); // Foreign Key
             $table->string('desc');
             $table->string('meta_title')->nullable();
             $table->string('meta_desc')->nullable();
